Added a setting to customize how the date formats appear (#121)
A number of users have contacted me, asking how to change the date format of the app to use their preferred setting. While such a capability exists within Android, it only works if you change your phone's language setting (e.g. from en-US to en-UK). This is a rather convoluted mechanism for making this work, so I added it as a setting instead.

The commit includes four date format options:

- Default. The phone's base version
- M_d_yyyy. The en-US style
- d_M_yyyy. The en-UK (and most of the rest of the world) style
- yyyy_MM_dd. For user's that wish to start with the year