Add a tooltip to explain how Smart Receipts reports/trips are suppose to function
A number of new users struggle to get started with the app and inevitable have questions about how to get started. While we have a variety of resources to assist with this (e.g. Video Guides), I'd like to do a better job of guiding their experience in an semi-intuitive manner.

To help explain how this works, I've added a tooltip to guide users on the new report screen.